Ignore editorconfig in etc/ and var/
These directories can be left alone to use whatever horrid style they
want, because they're not files I maintain.

This is necessary because of a nasty interaction between ido,
editorconfig, and no-littering, which was causing a "revert file"
prompt on shutdown.

I'm not sure of how this will be affected by .gitignore. But
whatever. I can always change that file if it doesn't clone/pull
cleanly.