Add new theme, uchū++
I love Snazzy, but I've been wanting a theme with higher contrast,
as well as one that is able to display the full 16 VGA colors.

This one is based on https://uchu.style/ - with a couple of extra
colors, one (magenta) from snazzy, the other (cyan/teal) I came up
with based on the others.